code

+++ Shopee project/shopee/src/main/java/fptu/shopee/Model/ListEmailToReceiveElectronicInvoices.java
@@ -11,18 +11,18 @@ public class ListEmailToReceiveElectronicInvoices {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "Id_List_Email_To_Receive_Electronic_Invoices")
-    private Long idListEmailToReceiveElectronicInvoices;
+    private int idListEmailToReceiveElectronicInvoices;
 
     @Email(message = "Email không hợp lệ.")
     @NotNull(message = "Email không được để trống.")
     @Column(name = "email", nullable = false)
     private String email;
 
-    public Long getIdListEmailToReceiveElectronicInvoices() {
+    public int getIdListEmailToReceiveElectronicInvoices() {
         return idListEmailToReceiveElectronicInvoices;
     }
 
-    public void setIdListEmailToReceiveElectronicInvoices(Long idListEmailToReceiveElectronicInvoices) {
+    public void setIdListEmailToReceiveElectronicInvoices(int idListEmailToReceiveElectronicInvoices) {
         this.idListEmailToReceiveElectronicInvoices = idListEmailToReceiveElectronicInvoices;
     }
